move TestDelEntryToJSON to entry_del_test.go
